feat: Build config-driven testing. Convert unit tests. (#38)
This creates a binary to run tests (and in the future, benchmarks) that are driven by config. After a few more changes (such as packaging the runner with Docker), this model will make it possible for plugin developers to run functional tests without any Bazel or C++ knowledge. The samples repo will continue to build plugins and invoke the runner using Bazel.

* Build config-driven testing. Convert unit tests.
* Generalize StringMatcher concept across logs + body.
* Fix typo

Signed-off-by: Martijn Stevenson <mstevenson@google.com>