feat: add support for running background streams in plugin benchmarks
Adds two command-line flags to the plugin tester:

--num_background_streams
    Specifies the number of background streams to instantiate in benchmarks.

--background_stream_advance_rate
    If background streams are configured, specifies how many background
    streams to advance the state of in each benchmark cycle.

Background streams are instantiated and advanced outside of the
benchmark timer. They pass the same headers and body inputs to the
plugin that are specified for the test case being benchmarked.

The num_background_streams and background_stream_advance_rate values can
also be configured by setting the corresponding fields of the
service_extensions.samples.pb.Env protobuf.

Signed-off-by: Michael Warres <mpw@google.com>