feat: add support for running background streams in plugin benchmarks (#248)
Adds two command-line flags to the plugin tester:

--num_additional_streams
Specifies the number of additional streams to instantiate in benchmarks.

--additional_stream_advance_rate
If additional streams are configured, specifies how many background
streams to advance the state of in each benchmark cycle.

Additional streams are instantiated and advanced outside of the benchmark timer. They pass the same headers and body inputs to the plugin that are specified for the test case being benchmarked.

Example usage:

```
$ cd plugins/
$ bazelisk build ...
$ bazel-bin/test/runner \
    --proto=samples/add_response_header/tests.textpb \
    --plugin=bazel-bin/samples/add_response_header/plugin_cpp.wasm \
    --num_additional_streams=1000 \
    --additional_stream_advance_rate=5
```

The num_additional_streams and additional_stream_advance_rate values can also be configured by setting the corresponding fields of the service_extensions.samples.pb.Env protobuf.

Fixes #191

Signed-off-by: Michael Warres <mpw@google.com>